requirements.txt
Flask==1.0.2
gunicorn==19.9.0
h5py==2.8.0
Jinja2==2.10
Keras==2.1.5
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.1.0
requests==2.19.1
scipy==1.0.1
tensorboard==1.7.0
tensorflow==1.7.0
tqdm==4.26.0
urllib3==1.22
utils==0.9.0
